Routine

+++ Gromacs/standard/pdb2gmx/mkgmx_ionized.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-rm -f editconf.pdb solvate.gro ions.tpr ionized.pdb
+MOL="$1"
+NOMOVE=true
+[ $# -eq 0 ] && { echo "mkvmd> Usage: $0 [MOLECULE]; Suggested: conf.gro"; exit 1; }
+
+if [ ! -f $MOL ]; then
+    echo "mkgmx> $MOL \nStarting system not found!"
+    exit 1
+fi
+
+rm editconf.gro solvate.gro ions.tpr ionized.gro
 
 cat > ions.mdp << EOF
 integrator  = md
@@ -19,8 +28,8 @@ EOF
 
 # Create a cubic box of 2 nm on each side and place the system to the center and align its principle axes to the reference axes
 # -princ usually helps to reduce your system size
-echo 0 | gmx editconf -f conf.gro -o editconf.gro -princ -d 1.0 -bt cubic
-#echo 0 | gmx editconf -f conf.gro -o editconf.gro -princ -center 3.7 3.4 2.4 -box 8.0 12.2 5.2
+echo 0 | gmx editconf -f $MOL -o editconf.gro -princ -d 1.0 -bt cubic
+#echo 0 | gmx editconf -f $MOL -o editconf.gro -princ -center 3.7 3.4 2.4 -box 8.0 12.2 5.2
 
 # Solvate the box
 gmx solvate -cp editconf.gro -o solvate.gro -p topol.top
@@ -30,4 +39,20 @@ gmx grompp -f ions.mdp -c solvate.gro -o ions.tpr -p topol.top -maxwarn 1 > LOG_
 gmx genion -s ions.tpr -o ionized.gro -conc 0.15 -neutral -p topol.top
 echo q | gmx make_ndx -f ionized.gro 
 
-rm -f \#* editconf.gro solvate.gro ions.mdp ions.tpr mdout.mdp
+if $NOMOVE; then
+# Add NOMOVE
+cp topol_Protein_chain_A.itp topol_Protein_chain_A.itp.BAK
+cat >> topol_Protein_chain_A.itp <<EOF
+
+#ifdef POSRES_NOMOVE
+#include "posre_Protein_chain_A_NOMOVE.itp"
+#endif
+EOF
+
+cat > posre_Protein_chain_A_NOMOVE.itp <<EOF
+[ position_restraints ]
+     3199     1  1000  1000  1000
+EOF
+fi
+
+rm \#* editconf.gro solvate.gro ions.tpr ions.mdp mdout.mdp
